+++ Changelog.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Reimplement the `Error` trait for `LocateManifestError`
+
 # 0.2.1 – 2020-08-30 (yanked)
 
 - Remove internal dependency on `thiserror` to reduce compile times